Add factories, modified models

+++ src/app/Services/InvoiceService.php
@@ -9,25 +9,69 @@ use App\Models\Customer;
 use App\Models\Fee;
 use App\Models\Invoice;
 use App\Models\Item;
+use Illuminate\Support\Str;
 
 class InvoiceService
 {
+    protected $company;
+    protected $customer;
+    protected $invoice;
+
     public function create($request)
     {
         try {
-            $invoice = $this->createInvoice($request);
-            $invoice->company = $this->createCompany($invoice, $request['company']);
-            $invoice->customer = $this->createCustomer($invoice, $request['customer']);
-            $invoice->items = $this->createItems($invoice, $request['items']);
-            $invoice->fees = $this->createFees($invoice, $request['fees']);
+            $this->company = $this->createCompany($request['company']);
+            $this->customer = $this->createCustomer($request['customer']);
+            $this->$invoice = $this->createInvoice($request);
+            $this->invoice->items = $this->createItems($request['items']);
+            $this->invoice->fees = $this->createFees($request['fees']);
 
-            return $invoice;
+            return $invoice->save();
         } catch (Exception $ex) {
-            Log::info($ex->getMessage());
+            Log::error($ex->getMessage());
             $this->rollback($request['id']);
         }
 
         return null;
+    }
+
+    private function createCompany($company)
+    {
+        $company = Company::create([
+            'id' => parseOrNull($company, 'id'),
+            'name' => parseOrNull($company, 'name'),
+            'street' => parseOrNull($company, 'street'),
+            'city' => parseOrNull($company, 'city'),
+            'stateProvince' => parseOrNull($company, 'stateProvince'),
+            'country' => parseOrNull($company, 'country'),
+            'zipPostal' => parseOrNull($company, 'zipPostal'),
+            'phone' => parseOrNull($company, 'phone'),
+            'email' => parseOrNull($company, 'email'),
+        ]);
+
+        return $company;
+    }
+
+    private function createCustomer($customer)
+    {
+        $customer = Customer::create([
+            'id' => parseOrNull($customer, 'id'),
+            'shippingName' => parseOrNull($customer, 'shippingName'),
+            'shippingStreet' => parseOrNull($customer, 'shippingStreet'),
+            'shippingCity' => parseOrNull($customer, 'shippingCity'),
+            'shippingStateProvince' => parseOrNull($customer, 'shippingStateProvince'),
+            'shippingCountry' => parseOrNull($customer, 'shippingCountry'),
+            'shippingZipPostal' => parseOrNull($customer, 'shippingZipPostal'),
+            'shippingPhone' => parseOrNull($customer, 'shippingPhone'),
+            'shippingEmail' => parseOrNull($customer, 'shippingEmail'),
+            'businessStreet' => parseOrNull($customer, 'businessStreet'),
+            'businessCity' => parseOrNull($customer, 'businessCity'),
+            'businessStateProvince' => parseOrNull($customer, 'businessStateProvince'),
+            'businessCountry' => parseOrNull($customer, 'businessCountry'),
+            'businessZipPostal' => parseOrNull($customer, 'businessZipPostal'),
+        ]);
+
+        return $customer;
     }
 
     private function createInvoice($invoice)
@@ -43,50 +87,14 @@ class InvoiceService
         return $invoice;
     }
 
-    private function createCompany($invoice, $company)
-    {
-        $company = Company::create([
-            'id' => parseOrNull($company, 'id'),
-            'invoice_id' => $invoice->id,
-            'name' => parseOrNull($company, 'name'),
-            'street' => parseOrNull($company, 'street'),
-            'city' => parseOrNull($company, 'city'),
-            'stateProvince' => parseOrNull($company, 'stateProvince'),
-            'country' => parseOrNull($company, 'country'),
-            'zipPostal' => parseOrNull($company, 'zipPostal'),
-            'phone' => parseOrNull($company, 'phone'),
-            'email' => parseOrNull($company, 'email'),
-        ]);
-
-        return $company;
-    }
-
-    private function createCustomer($invoice, $customer)
-    {
-        $customer = Customer::create([
-            'id' => parseOrNull($customer, 'id'),
-            'invoice_id' => $invoice->id,
-            'name' => parseOrNull($customer, 'name'),
-            'street' => parseOrNull($customer, 'street'),
-            'city' => parseOrNull($customer, 'city'),
-            'stateProvince' => parseOrNull($customer, 'stateProvince'),
-            'country' => parseOrNull($customer, 'country'),
-            'zipPostal' => parseOrNull($customer, 'zipPostal'),
-            'phone' => parseOrNull($customer, 'phone'),
-            'email' => parseOrNull($customer, 'email'),
-        ]);
-
-        return $customer;
-    }
-
-    private function createItems($invoice, $items)
+    private function createItems($items)
     {
         $newItems = [];
 
         foreach ($items as $item) {
             $newItems[] = Item::create([
                 'id' => parseOrNull($item, 'id'),
-                'invoice_id' => $invoice->id,
+                'invoice_id' => $this->invoice->id,
                 'description' => parseOrNull($item, 'description'),
                 'units' => parseOrNull($item, 'units'),
                 'pricePerUnit' => parseOrNull($item, 'pricePerUnit'),
@@ -97,14 +105,14 @@ class InvoiceService
         return collect($newItems);
     }
 
-    private function createFees($invoice, $fees)
+    private function createFees($fees)
     {
         $newFees = [];
 
         foreach ($fees as $fee) {
             $newFees[] = Fee::create([
                 'id' => parseOrNull($fee, 'id'),
-                'invoice_id' => $invoice->id,
+                'invoice_id' => $this->invoice->id,
                 'description' => parseOrNull($fee, 'description'),
                 'value' => parseOrNull($fee, 'value'),
                 'type' => parseOrNull($fee, 'type'),
@@ -117,10 +125,18 @@ class InvoiceService
 
     private function rollback($invoiceId)
     {
-        Invoice::where('id', $invoiceId)->delete();
-        Company::where('invoice_id', $invoiceId)->delete();
-        Customer::where('invoice_id', $invoiceId)->delete();
-        Item::where('invoice_id', $invoiceId)->delete();
-        Fee::where('invoice_id', $invoiceId)->delete();
+        if ($this->company) {
+            Company::where('invoice_id', $invoiceId)->delete();
+        }
+
+        if ($this->customer) {
+            Customer::where('invoice_id', $invoiceId)->delete();
+        }
+
+        if ($this->invoice) {
+            Invoice::where('id', $invoiceId)->delete();
+            Item::where('invoice_id', $invoiceId)->delete();
+            Fee::where('invoice_id', $invoiceId)->delete();
+        }
     }
 }
